backend/requirements.txt
fastapi
uvicorn
requests
pymupdf
python-multipart
langchain-community
langchain-chroma
langchain-openai
openai
python-dotenv
pydantic
aiofiles
pydantic[email]
scikit-learn
python-docx
mime
cryptography
unstructured
pandas
ollama
